# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trackvance Core",
    "description": "Monolito modular local. SUCCESS es estado técnico; decision conserva el resultado de negocio. Configuraciones y evidence son snapshots versionados.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/monitors/{monitor_id}/schedule": {
      "get": {
        "tags": [
          "Sentinel"
        ],
        "summary": "Get Schedule",
        "operationId": "get_schedule_api_v1_monitors__monitor_id__schedule_get",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monitor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ScheduleResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Schedule Api V1 Monitors  Monitor Id  Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "runs:read"
      },
      "post": {
        "tags": [
          "Sentinel"
        ],
        "summary": "Configure Schedule",
        "operationId": "configure_schedule_api_v1_monitors__monitor_id__schedule_post",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monitor Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "configurations:write"
      }
    },
    "/api/v1/monitors/{monitor_id}/occurrences": {
      "get": {
        "tags": [
          "Sentinel"
        ],
        "summary": "Occurrences",
        "operationId": "occurrences_api_v1_monitors__monitor_id__occurrences_get",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monitor Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "runs:read"
      }
    },
    "/api/v1/monitors/{monitor_id}/series": {
      "get": {
        "tags": [
          "Sentinel"
        ],
        "summary": "Metric Series",
        "operationId": "metric_series_api_v1_monitors__monitor_id__series_get",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monitor Id"
            }
          },
          {
            "name": "include_versions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Versions"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricSeriesCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "runs:read"
      }
    },
    "/api/v1/monitors/{monitor_id}/alerts": {
      "get": {
        "tags": [
          "Sentinel"
        ],
        "summary": "Monitor Alerts",
        "operationId": "monitor_alerts_api_v1_monitors__monitor_id__alerts_get",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monitor Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "runs:read"
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetricPointResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "configuration_id": {
            "type": "string",
            "title": "Configuration Id"
          },
          "dataset_version_id": {
            "type": "string",
            "title": "Dataset Version Id"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "exact_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact Value"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "configuration_id",
          "dataset_version_id",
          "observed_at",
          "value",
          "exact_value",
          "decision"
        ],
        "title": "MetricPointResponse"
      },
      "MetricSeriesCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MetricSeriesResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "includes_configuration_versions": {
            "type": "boolean",
            "title": "Includes Configuration Versions"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "sample_count",
          "limit",
          "includes_configuration_versions"
        ],
        "title": "MetricSeriesCollection"
      },
      "MetricSeriesResponse": {
        "properties": {
          "metric_key": {
            "type": "string",
            "title": "Metric Key"
          },
          "dimensions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dimensions"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "metric_definition_version": {
            "type": "integer",
            "title": "Metric Definition Version"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/MetricPointResponse"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "metric_key",
          "dimensions",
          "method",
          "metric_definition_version",
          "points"
        ],
        "title": "MetricSeriesResponse"
      },
...
      "OccurrenceCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OccurrenceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "OccurrenceCollection"
      },
      "OccurrenceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "schedule_id": {
            "type": "string",
            "title": "Schedule Id"
          },
          "schedule_version_id": {
            "type": "string",
            "title": "Schedule Version Id"
          },
          "monitor_id": {
            "type": "string",
            "title": "Monitor Id"
          },
          "dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "planned_at": {
            "type": "string",
            "title": "Planned At"
          },
          "dispatched_at": {
            "type": "string",
            "title": "Dispatched At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "coalesced_intervals": {
            "type": "integer",
            "title": "Coalesced Intervals"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "schedule_id",
          "schedule_version_id",
          "monitor_id",
          "dataset_version_id",
          "run_id",
          "planned_at",
          "dispatched_at",
          "started_at",
          "finished_at",
          "status",
          "decision",
          "reason_code",
          "coalesced_intervals",
          "metrics"
        ],
        "title": "OccurrenceResponse"
      },
...
      "ScheduleBody": {
        "properties": {
          "interval_seconds": {
            "type": "integer",
            "maximum": 2678400.0,
            "minimum": 60.0,
            "title": "Interval Seconds"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "interval_seconds"
        ],
        "title": "ScheduleBody"
      },
      "ScheduleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "monitor_id": {
            "type": "string",
            "title": "Monitor Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "schedule_version_id": {
            "type": "string",
            "title": "Schedule Version Id"
          },
          "interval_seconds": {
            "type": "integer",
            "title": "Interval Seconds"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "next_run_at": {
            "type": "string",
            "title": "Next Run At"
          },
          "starts_at": {
            "type": "string",
            "title": "Starts At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "source_policy": {
            "type": "string",
            "title": "Source Policy"
          },
          "misfire_policy": {
            "type": "string",
            "title": "Misfire Policy"
          },
          "overlap_policy": {
            "type": "string",
            "title": "Overlap Policy"
          }
        },
        "type": "object",
        "required": [
          "id",
          "monitor_id",
          "version",
          "schedule_version_id",
          "interval_seconds",
          "enabled",
          "next_run_at",
          "starts_at",
          "updated_at",
          "source_policy",
          "misfire_policy",
          "overlap_policy"
        ],
        "title": "ScheduleResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "LocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "trackvance_session",
        "description": "Sesión local HttpOnly. Las mutaciones requieren además X-CSRF-Token."
      }
    }
  }
}